requirements.txt
python-telegram-bot
tensorflow>=2.12.0
numpy>=1.23.5
opencv-python>=4.7.0.72
